requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas
xlrd==1.2.0
openpyxl==3.0.3
azure-storage-blob==12.6.0
# cryptography==2.8